% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult.R
\docType{data}
\name{adult}
\alias{adult}
\title{Adult dataset}
\format{
A data frame of 48842 rows and 15 columns
\describe{
 \item{age}{Numeric values}
 \item{workclass}{A factor with 8 levels}
 \item{fnlwgt}{Numeric values}
 \item{education}{}
 \item{education_num}{}
 \item{marital_status}{}
 \item{occupation}{}
 \item{relationship }{}
 \item{race}{}
 \item{sex}{}
 \item{capital_gain}{}
 \item{capital_loss}{}
 \item{hours_per_week}{}
 \item{native_country}{}
 \item{class_labels}{}
}
}
\source{
... data
}
\usage{
adult
}
\description{
A dataset containing ... data.
}
\keyword{datasets}
