% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNA.R
\name{createNA}
\alias{createNA}
\title{Create missing values for a dataset}
\usage{
createNA(data, names = NULL, p = 0.3, seed = NULL)
}
\arguments{
\item{data}{A complete data frame.}

\item{names}{The names of variables where missing values will be generated.}

\item{p}{The proportion of missing values in the data frame or the proportions of missing values corresponding to the variables specified in \code{names}.}

\item{seed}{An integer value. Default: \code{NULL}}
}
\value{
A data frame with artificial missing values
}
\description{
This function creates missing values under the missing complete at random mechanism (MCAR). It is for demonstration purposes only.
}
\examples{
# Create 30\% MCAR data across all variables in a dataset
withNA.df <- createNA(iris, p = 0.3)

# Create 30\% MCAR data in a specified variable in a dataset
withNA.df <- createNA(iris, names = c("Sepal.Length"), p = 0.3)

# Create MCAR data in several specified variables in a dataset
withNA.df <- createNA(iris, names = c("Sepal.Length", "Petal.Width", "Species"),
                      p = c(0.3, 0.2, 0.1))
}
