% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{feature_type}
\alias{feature_type}
\title{Return the type ("numeric", "integer","binary","multiclass") of each variable in the dataset}
\usage{
feature_type(data)
}
\arguments{
\item{data}{A data.frame or a data.table}
}
\description{
Return the type ("numeric", "integer","binary","multiclass") of each variable in the dataset
}
