% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_impnew.R
\name{imp.mean2}
\alias{imp.mean2}
\title{impute the missing values of a vector with the mean of observed values}
\usage{
imp.mean2(vec, na.idx = NULL, traindata.vec)
}
\arguments{
\item{vec}{a vector of numeric values}

\item{na.idx}{indices of missing values}

\item{traindata.vec}{the corresponding vector in the train set}

\item{obs.idx}{indices of observed values}
}
\description{
impute the missing values of a vector with the mean of observed values
}
