% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_imp.R
\name{imp.median}
\alias{imp.median}
\title{impute the missing values of a vector with the median of observed values}
\usage{
imp.median(vec, obs.idx = NULL, na.idx = NULL)
}
\arguments{
\item{vec}{a vector of numeric values}

\item{na.idx}{indices of missing values}
}
\description{
impute the missing values of a vector with the median of observed values
}
