% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_impnew.R
\name{imp.median2}
\alias{imp.median2}
\title{impute the missing values of a vector with the median of observed values}
\usage{
imp.median2(vec, na.idx = NULL, traindata.vec)
}
\arguments{
\item{vec}{a vector of numeric values}

\item{na.idx}{indices of missing values}

\item{traindata.vec}{the corresponding vector in the train set}
}
\description{
impute the missing values of a vector with the median of observed values
}
