% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_imp.R
\name{imp.mode}
\alias{imp.mode}
\title{impute the missing values of a vector with the mode (majority class) of observed values}
\usage{
imp.mode(vec, obs.idx = NULL, na.idx = NULL)
}
\arguments{
\item{vec}{a vector of numeric values (ideally integer type) or factor}

\item{na.idx}{indices of missing values}
}
\description{
impute the missing values of a vector with the mode (majority class) of observed values
}
