% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_impnew.R
\name{imp.mode2}
\alias{imp.mode2}
\title{impute the missing values of a vector with the mode (majority class) of observed values}
\usage{
imp.mode2(vec, na.idx = NULL, traindata.vec)
}
\arguments{
\item{vec}{a vector of numeric values (ideally integer type) or factor}

\item{na.idx}{indices of missing values}

\item{traindata.vec}{the corresponding vector in the train set}
}
\description{
impute the missing values of a vector with the mode (majority class) of observed values
}
