% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_imp.R
\name{imp.normal}
\alias{imp.normal}
\title{impute the missing values of a vector with randomly selected values from a normal distribution with mean and sd extracted from observed values}
\usage{
imp.normal(vec, na.idx = NULL)
}
\arguments{
\item{vec}{a vector of numeric values}

\item{na.idx}{indices of missing values}
}
\description{
impute the missing values of a vector with randomly selected values from a normal distribution with mean and sd extracted from observed values
}
