% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_impnew.R
\name{imp.normal2}
\alias{imp.normal2}
\title{impute the missing values of a vector with randomly selected values from a normal distribution with mean and sd extracted from observed values}
\usage{
imp.normal2(vec, na.idx = NULL, traindata.vec)
}
\arguments{
\item{vec}{a vector of numeric values}

\item{na.idx}{indices of missing values}

\item{traindata.vec}{the corresponding vector in the train set}
}
\description{
impute the missing values of a vector with randomly selected values from a normal distribution with mean and sd extracted from observed values
}
