% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_imp.R
\name{imp.sample}
\alias{imp.sample}
\title{impute the missing values of a vector with sampled observed values}
\usage{
imp.sample(vec, na.idx = NULL)
}
\arguments{
\item{vec}{a vector of numeric or factor values}

\item{na.idx}{indices of missing values}
}
\description{
impute the missing values of a vector with sampled observed values
}
