% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_impnew.R
\name{imp.sample2}
\alias{imp.sample2}
\title{impute the missing values of a vector with sampled observed values}
\usage{
imp.sample2(vec, na.idx = NULL, traindata.vec)
}
\arguments{
\item{vec}{a vector of numeric or factor values in the newdata set}

\item{na.idx}{indices of missing values}

\item{traindata.vec}{the corresponding vector in the train set}
}
\description{
impute the missing values of a vector with sampled observed values
}
