% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-new.R
\name{impute.new}
\alias{impute.new}
\title{This function is used to impute new data using training imputer models}
\usage{
impute.new(object, newdata, pmm.new = FALSE, pmm.k = NULL)
}
\arguments{
\item{object}{training imputer object}

\item{newdata}{A data frame}
}
\description{
This function is used to impute new data using training imputer models
}
