% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_new.R
\name{impute_new}
\alias{impute_new}
\title{Impute new data using saved imputation models from training data}
\usage{
impute_new(object, newdata, initial.newdata = FALSE, pmm.k = NULL, m = NULL)
}
\arguments{
\item{object}{training imputer object}

\item{newdata}{a data.frame or data.table}

\item{initial.newdata}{whether or not to use the information of the new data to initially impute new data (mean,median,sd etc). Default: FALSE (use training set information instead)}

\item{pmm.k}{the number of donors for predictive mmean matching. Default: NULL}

\item{m}{the number of imputed datasets. Default: NULL (m would be set to the same value as the training imputer object)}
}
\description{
Impute new data using saved imputation models from training data
}
