% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{imputed_values}
\alias{imputed_values}
\title{Use to return a data.table with imputed values of a variable with missing values}
\usage{
imputed_values(imputation.list, var.name, na.idx, true.values = NULL)
}
\arguments{
\item{imputation.list}{a list of m imputed dataset return by mixgb imputer}

\item{var.name}{the name of a variable of interest}

\item{na.idx}{the indices of the missing values of this variable}
}
\value{
a data.table with m columns, each columns represent the imputed values of the missing values in this variable
}
\description{
Use to return a data.table with imputed values of a variable with missing values
}
