% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_imp.R
\name{initial_imp}
\alias{initial_imp}
\title{This function is use to initially impute missing value and validate the feed-in dataset}
\usage{
initial_imp(
  data,
  initial.num = "normal",
  initial.fac = "mode",
  bootstrap = TRUE
)
}
\arguments{
\item{data}{A data table (with missing values NA's)}

\item{initial.num}{initial imputation method for numeric type data ("normal","mean","median","mode","sample"). Default: "normal"}

\item{initial.fac}{initial imputation method for factor type data ("mode","sample"). Default: "mode"}

\item{bootstrap}{whether or not use bootstrap for multiple imputation. If TRUE, also return sortedNA.dt}
}
\value{
A list of items including data table with initial imputation, sorted index, Types, Names etc.
}
\description{
This function is use to initially impute missing value and validate the feed-in dataset
}
