% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_impnew.R
\name{initial_impnew}
\alias{initial_impnew}
\title{This function is use to initially impute missing value in the new data using information of training data and validate the feed-in dataset}
\usage{
initial_impnew(
  initial.newdata = FALSE,
  new.sorted,
  traindata,
  sorted.types,
  sorted.names,
  initial.num = "normal",
  initial.fac = "mode",
  bootstrap = TRUE
)
}
\arguments{
\item{initial.newdata}{whether or not to use the information of the new data to initially impute new data (mean,median,sd etc). Default: FALSE (use training set information instead)}

\item{new.sorted}{A data.table (with missing values NA's). Must be of the same order as sorted training data.}

\item{traindata}{A data.table}

\item{initial.num}{initial imputation method for numeric type data ("normal","mean","median","mode","sample"). Default: "normal"}

\item{initial.fac}{initial imputation method for factor type data ("mode","sample"). Default: "mode"}

\item{bootstrap}{whether or not use bootstrap for multiple imputation. If TRUE, also return sortedNA.dt}
}
\value{
A list of items including data table with initial imputation, sorted index, Types, Names etc.
}
\description{
This function is use to initially impute missing value in the new data using information of training data and validate the feed-in dataset
}
