% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixgb_boot.R
\name{mixgb_boot}
\alias{mixgb_boot}
\title{multiple imputation using xgboost with bootstrap}
\usage{
mixgb_boot(
  BNa.idx,
  boot.dt,
  pmm.type,
  pmm.link,
  pmm.k,
  yobs.list,
  yhatobs.list,
  sorted.dt,
  missing.vars,
  sorted.names,
  Na.idx,
  missing.types,
  Ncol,
  xgb.params = list(max_depth = 6, gamma = 0.1, eta = 0.3, colsample_bytree = 1,
    min_child_weight = 1, subsample = 1, tree_method = "auto", gpu_id = 0, predictor =
    "auto", scale_pos_weight = 1),
  nrounds = 50,
  early_stopping_rounds = 10,
  print_every_n = 10L,
  verbose = 0,
  ...
)
}
\description{
multiple imputation using xgboost with bootstrap
}
