% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixgb_cv.R
\name{mixgb_cv}
\alias{mixgb_cv}
\title{Use cross-validation to find the optimum nrounds}
\usage{
mixgb_cv(
  data,
  nfold = 5,
  nrounds = 100,
  early_stopping_rounds = 10,
  response = NULL,
  select_features = NULL,
  stringsAsFactors = FALSE,
  ...
)
}
\arguments{
\item{data}{a data.frame or a data.table with missing values}

\item{select_features}{names or indices of the selected features. Default: NULL (select all features)}

\item{reponse}{name or column index of the response variable. Default: NULL.}
}
\description{
Use cross-validation to find the optimum nrounds
}
