% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixgb_save.R
\name{mixgb_save}
\alias{mixgb_save}
\title{multiple imputation using xgboost (save models and imputations)}
\usage{
mixgb_save(
  save.vars,
  save.p,
  extra.vars = NULL,
  extra.types = NULL,
  pmm.type,
  pmm.link,
  pmm.k,
  yobs.list,
  yhatobs.list = NULL,
  sorted.dt,
  missing.vars,
  sorted.names,
  Na.idx,
  missing.types,
  Ncol,
  xgb.params = list(max_depth = 6, gamma = 0.1, eta = 0.3, colsample_bytree = 1,
    min_child_weight = 1, subsample = 1, tree_method = "auto", gpu_id = 0, predictor =
    "auto", scale_pos_weight = 1),
  nrounds = 50,
  early_stopping_rounds = 10,
  print_every_n = 10L,
  verbose = 0,
  ...
)
}
\arguments{
\item{yhatobs.list}{if it is pmm.type 1, must feed in the yhatobs.list}
}
\description{
multiple imputation using xgboost (save models and imputations)
}
