% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixgb_use.R
\name{mixgb_use}
\alias{mixgb_use}
\title{multiple imputation using xgboost (through saved models)}
\usage{
mixgb_use(
  m.set,
  xgb.models,
  save.vars,
  save.p,
  extra.vars = NULL,
  extra.types = NULL,
  pmm.type,
  pmm.link,
  pmm.k,
  yobs.list,
  yhatobs.list = NULL,
  sorted.dt,
  missing.vars,
  sorted.names,
  Na.idx,
  missing.types,
  Ncol
)
}
\arguments{
\item{m.set}{the ith imputation}

\item{yobs.list}{observed values in the original training data}

\item{yhatobs.list}{predicted observed values in the original training data,}

\item{sorted.dt}{sorted new data after initial imputation (with the same order as the original data)}

\item{missing.vars}{names of variables with missing values in new data}

\item{sorted.names}{all names of variables in sorted order}

\item{missing.types}{types of variable with missing values in new data}

\item{Ncol}{number of columns in new data}
}
\description{
multiple imputation using xgboost (through saved models)
}
