% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3vars.R
\name{plot_1num2fac}
\alias{plot_1num2fac}
\title{Box plots with overlaying data points for a numeric variable vs a factor condition on another factor}
\usage{
plot_1num2fac(
  imputation.list,
  var.fac,
  var.num,
  con.fac,
  original.data,
  true.data = NULL,
  color.pal = NULL,
  shape.off = TRUE
)
}
\arguments{
\item{imputation.list}{A list of \code{m} imputed datasets returned by the \code{mixgb} imputer}

\item{var.fac}{A factor variable on the x-axis}

\item{var.num}{A numeric variable on the y-axis}

\item{con.fac}{A conditional factor}

\item{original.data}{The original data with missing data}

\item{true.data}{The true data without missing values. In general, this is unknown. Only use for simulation studies.}

\item{color.pal}{A vector of hex color codes for the observed and m sets of imputed values panels. The vector should be of length \code{m+1}. Default: NULL (use "gray40" for the observed panel, use ggplot2 default colors for other panels.)}

\item{shape.off}{Whether to turn off the shape symbols for plotting. By default, this is set to TRUE to speed up plotting. We only recommend using `shape.off = FALSE` for small datasets.}
}
\value{
Boxplots with overlaying data points
}
\description{
Plot observed values versus m sets of imputed values for one specified numeric variable and two factors using \pkg{ggplot2}.
}
\examples{
\donttest{

#obtain m multiply datasets
imputed.data <- mixgb(data = nhanes3_newborn, m = 5)

#plot the multiply imputed values for variables "BMPRECUM" versus "HFF1" conditional on "HSSEX"
plot_1num2fac(
  imputation.list = imputed.data, var.fac = "HFF1", var.num = "BMPRECUM",
  con.fac = "HSSEX", original.data = nhanes3_newborn)
}
}
