% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_1var.R
\name{plot_box}
\alias{plot_box}
\title{Boxplots with data points for multiply imputed values for a single numeric variable}
\usage{
plot_box(imputation.list, var.num, original.data, color.pal = NULL)
}
\arguments{
\item{imputation.list}{A list of \code{m} imputed datasets}

\item{var.num}{The name of a numeric variable of interest}

\item{original.data}{The original data with missing data}

\item{color.pal}{A vector of hex color codes for the observed and m sets of imputed values panels. The vector should be of length \code{m+1}. Default: NULL (use "gray40" for the observed panel, use ggplot2 default colors for other panels.)}
}
\value{
Boxplots with data points for a numeric variable
}
\description{
Plot boxplots with data points of observed values versus m sets of imputed values for a specified numeric variable using \pkg{ggplot2}.
}
\examples{
\donttest{

#obtain m multiply datasets
imputed.data <- mixgb(data = nhanes3_newborn, m = 5)

#plot the multiply imputed values for variable "BMPHEAD"
plot_box(imputation.list = imputed.data, var.num = "BMPHEAD",
  original.data = nhanes3_newborn)
}
}
