% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm.R
\name{pmm.multiclass}
\alias{pmm.multiclass}
\title{PMM for multiclass variable}
\usage{
pmm.multiclass(yhatobs, yhatmis, yobs, k)
}
\arguments{
\item{yhatobs}{The predicted values of observed entries in a variable}

\item{yhatmis}{The predicted values of missing entries in a variable}

\item{yobs}{The actual observed values of observed entries in a variable}

\item{k}{The number of donors.}
}
\value{
The matched observed values of all missing entries
}
\description{
PMM for multiclass variable
}
\keyword{internal}
