% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_vars.R
\name{save_vars}
\alias{save_vars}
\title{checking save.vars}
\usage{
save_vars(save.vars, origin.names, missing.vars)
}
\value{
extra.vars, the names of the extra variables need to be saved models for.
}
\description{
checking save.vars
}
\examples{
origin.names <- c("age", "bmi", "hyp", "chl")
save_vars(save.vars = c(1, 3), origin.names = colnames(nhanes2), missing.vars = c("hyp", "bmi", "chl"))
save_vars(save.vars = c(1, 2, 3), origin.names = colnames(nhanes2), missing.vars = c("hyp", "bmi", "chl"))
save_vars(save.vars = c(2, 3, 4), origin.names = colnames(nhanes2), missing.vars = c("hyp", "bmi", "chl"))
save_vars(save.vars = c(1, 2, 3, 4), origin.names = colnames(nhanes2), missing.vars = c("hyp", "bmi", "chl"))
}
