% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_yhatobs.R
\name{save_yhatobs}
\alias{save_yhatobs}
\title{A function use to obtain yhatobs using the whole dataset (use for pmm.type=1)}
\usage{
save_yhatobs(
  yobs.list,
  maxit,
  pmm.link,
  sorted.dt,
  missing.vars,
  extra.vars = NULL,
  extra.types = NULL,
  sorted.names,
  Na.idx,
  missing.types,
  Ncol,
  xgb.params = list(max_depth = 6, gamma = 0.1, eta = 0.3, colsample_bytree = 1,
    min_child_weight = 1, subsample = 1, tree_method = "auto", gpu_id = 0, predictor =
    "auto", scale_pos_weight = 1),
  nrounds = 50,
  early_stopping_rounds = 10,
  print_every_n = 10L,
  verbose = 0,
  ...
)
}
\arguments{
\item{xgb.params}{NULL if XGBmodels was fed in}
}
\value{
a list of yhatobs values for specified variables
}
\description{
A function use to obtain yhatobs using the whole dataset (use for pmm.type=1)
}
