% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{sortNA}
\alias{sortNA}
\title{Sort data by increasing number of missing values}
\usage{
sortNA(data)
}
\arguments{
\item{data}{A data table (with missing values NA's)}
}
\description{
Sort data by increasing number of missing values
}
