% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{sortNA_df}
\alias{sortNA_df}
\title{Sort the dataset by increasing number of missing values (use in old version Xgb-imputer0)}
\usage{
sortNA_df(data)
}
\arguments{
\item{data}{A data frame (with missing values NA's)}
}
\description{
Sort the dataset by increasing number of missing values (use in old version Xgb-imputer0)
}
