% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{variable_class}
\alias{variable_class}
\title{This function is used to obtain a list of binary and multiclass variables}
\usage{
variable_class(data)
}
\arguments{
\item{data}{A data frame}
}
\description{
This function is used to obtain a list of binary and multiclass variables
}
