% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Data cleaning}
\usage{
check_data(data, verbose = TRUE, max_levels = 50, stringsAsFactors = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{verbose}{Verbose setting. If \code{TRUE}, will print out warnings when data issues are found. Default: \code{TRUE}.}

\item{max_levels}{An integer specifying the maximum number of levels allowed for factor variables. Default: 50.}

\item{stringsAsFactors}{A logical value indicating whether all character vectors in the dataset should be converted to factors. Default: \code{TRUE}.}
}
\value{
A preliminary cleaned dataset
}
\description{
The function `check_data()` serves the purpose of performing a preliminary check and fix some evident issues. However, the function cannot resolve all data quality-related problems.
}
\examples{
nhanes3[4, 4] <- NaN
nhanes3[5, 5] <- Inf
nhanes3[6, 6] <- -Inf

cleandata <- check_data(data = nhanes3)
}
