% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNA.R
\name{createNA}
\alias{createNA}
\title{Create missing value for a dataset}
\usage{
createNA(data, names = NULL, p = 0.3, seed = NULL)
}
\arguments{
\item{data}{a dataframe}

\item{p}{the proportion of missing values in the dataframe}

\item{seed}{random seed, a singlue value}
}
\description{
Create missing value for a dataset
}
\examples{
#create 30\% MCAR data across all variables in a dataset
withNA.df<-createNA(iris,p=0.3)

#create 30\% MCAR data in a specified variable in a dataset
withNA.df=createNA(iris,names=c("Sepal.Length"),p=0.3)

#create MCAR data in several specified variables in a dataset
withNA.df=createNA(iris,names=c("Sepal.Length","Petal.Width","Species"),p=c(0.3,0.2,0.1))
}
