% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocess.R
\name{feature_type}
\alias{feature_type}
\title{Return the type of each variable in the dataset}
\usage{
feature_type(data)
}
\arguments{
\item{data}{A data frame}
}
\description{
Return the type of each variable in the dataset
}
