% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-new.R
\name{impute.new}
\alias{impute.new}
\title{This function is used to impute new data using training imputer models}
\usage{
impute.new(object, newdata, pmm.new = FALSE, pmm.k = NULL, m = NULL)
}
\arguments{
\item{object}{training imputer object}

\item{newdata}{a data frame}

\item{pmm.new}{whether or not to apply predictive mean matching with the new dataset. Default: FALSE}

\item{pmm.k}{the number of donors for predictive mmean matching. Default: NULL}

\item{m}{the number of imputed datasets. Default: NULL}
}
\description{
This function is used to impute new data using training imputer models
}
