% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixgb-package.R
\docType{package}
\name{mixgb-package}
\alias{mixgb-package}
\title{mixgb: Multiple Imputation Through 'XGBoost'}
\description{
Multiple imputation using 'XGBoost', subsampling, and predictive mean matching as described in Deng and Lumley (2023) \doi{10.1080/10618600.2023.2252501}. The package supports various types of variables, offers flexible settings, and enables saving an imputation model to impute new data. Data processing and memory usage have been optimised to speed up the imputation process.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/agnesdeng/mixgb}
  \item Report bugs at \url{https://github.com/agnesdeng/mixgb/issues}
}

}
\author{
\strong{Maintainer}: Yongshi Deng \email{agnes.yongshideng@gmail.com} (\href{https://orcid.org/0000-0001-5845-859X}{ORCID})

Other contributors:
\itemize{
  \item Thomas Lumley \email{t.lumley@auckland.ac.nz} [thesis advisor]
}

}
\keyword{internal}
