% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-mixgb.R
\docType{package}
\name{mixgb}
\alias{mixgb}
\alias{mixgb-package}
\title{mixgb: multiple imputation through XGBoost}
\description{
Generate multiple imputations for large datasets using machine learning techniques.
}
\author{
\strong{Maintainer}: Yongshi Deng \email{yongshi.deng@auckland.ac.nz}

Other contributors:
\itemize{
  \item Thomas Lumley \email{t.lumley@auckland.ac.nz} [thesis advisor]
}

}
