% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocess.R
\name{sortNA}
\alias{sortNA}
\title{Sort the dataset by increasing number of missing values}
\usage{
sortNA(data)
}
\arguments{
\item{data}{A data frame (with missing values NA's)}
}
\description{
Sort the dataset by increasing number of missing values
}
